Implement first version of Dijkstra
Apart from the implementation of the function, some other modifications were made for it to work.
I modified a lot the DijkstraService class:
- Deleted the old method compareTo
- Created the classes Node, CellInfo and TableUltime and the functions that went with them.
- Modified computeShortestPath so that it gives Dijkstra what it needs.

I also made a small modification to GrapheComplet so that it is initialised with the max value instead of -1, and so that the costs were double instead of int.