Modified DijkstraService to make it handle more things and be more efficient
Modified DijkstraService so that :
- it stores map and DijkstraTable but not graph
- it initialises DijkstraTable upon construction
- it computes the shortest path, it takes the list of intersectionIDs to visit and returns the graph containing the shortest paths
- it reuses the values already calculated for every new run of computeShortestPath
- computeShortestPath and dijkstra return the graph

Also modified the Tests for DijkstraService to adapt to the changes and to suppress duplicates of the same tests.

Also modified PlanningService to adapt to the changes of DijkstraService. Dijkstra service is now a persisted attribute in PlanningService

modified PlannningService and TourService to handle different stops with same intersection. added deparsers, and modified generate precedences output, as well as graphToTour input. Documented code, implemented complementary tests.